date

+++ utils/formatDate.js
@@ -1,4 +1,4 @@
 export const formatDate = (date) => {
-  const options = { year: 'numeric', month: 'long', day: 'numeric' }
+  const options = { year: 'numeric', month: 'numeric', day: 'numeric' }
   return new Date(date).toLocaleDateString('cs', options)
 }
